backend/requirements.txt
# Web Framework & Server
fastapi==0.115.8
uvicorn[standard]==0.34.0

# Data Validation & Settings (Pydantic v2)
pydantic==2.10.6
pydantic-settings==2.7.1

# Database Drivers
asyncpg==0.30.0
psycopg2-binary==2.9.10
pyodbc>=5.2.0

# AI & LLM Clients
openai==1.61.0
anthropic==0.45.0

# Utilities
python-dotenv==1.0.1
httpx>=0.27.0,<0.28.0
fastmcp==0.4.1